move buttons + userControls
