+++ src/components/UIElements/UserControls/UserControls.jsx
@@ -1,4 +1,5 @@
 import Buttons from "../Buttons/Buttons";
+
 const lang = ["EN", "UA", "RU"];
 
 function UserControls() {
